requirements.txt
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.1
Flask-Migrate==2.0.3
alembic==0.9.1
PyJWT
gunicorn
psycopg2
sqlalchemy-utils
passlib
requests
pusher
lxml
elastic-apm[flask]
gevent
python-dotenv
ipdb
